requirements.txt
Crwy>=1.0.5
Django>=1.10.5
django-filter>=1.0.1
django-suit>=0.2.23
djangorestframework>=3.5.3
docutils>=0.13.1
redis>=2.10.5
SQLAlchemy>=1.1.4
